+++ 15 - Chiton/astar.cc
@@ -23,12 +23,12 @@ struct Vec2 {
 
 template <> struct std::hash<Vec2> {
   std::size_t operator()(const Vec2 &v) const noexcept {
-    return std::hash<uint64_t>{}((v.x << 16 | v.x >> 48) ^ v.y);
+    return std::hash<int64_t>{}((v.x << 16 | v.x >> 48) ^ v.y);
   }
 };
 
-int64_t manhattenDist(const Vec2 &a, const Vec2 &b) {
-  return std::abs(a.x - b.x) + std::abs(a.y - b.y);
+uint64_t manhattenDist(const Vec2 &a, const Vec2 &b) {
+  return static_cast<uint64_t>(std::abs(a.x - b.x) + std::abs(a.y - b.y));
 }
 
 template <typename FieldType> class Map {
@@ -49,7 +49,10 @@ public:
 
   size_t getWidth() const { return width - 1; }
   size_t getHeight() const { return height - 1; }
-  auto &operator[](const Vec2 &pos) { return map.at(pos.x + pos.y * width); }
+  auto &operator[](const Vec2 &pos) {
+    return map.at(
+        static_cast<size_t>(pos.x + pos.y * static_cast<int64_t>(width)));
+  }
 
   bool validCoords(const Vec2 &coord) {
     return (coord.x > 0 && std::cmp_less(coord.x, width - 1)) &&
